Remove logstash library
The requirements file installs both the `logstash` and `python-logstash`
libraries. The former isn't compatible with Python 3 (the app's requirement)
and conflicts with the latter. The only location that logstash appears to be
used (`dataactcore/utils/cloudLogger.py`), appears to be assuming the latter
library.

Combined, it seems like a safe bet that `logstash` isn't necessary and only
causes problems with new installs.