Refactor FSRS library to account for two services
Previously, the FSRS code assumed that it'd be grabbing from only one
location. In fact, there are two endpoints, one serving Procurements/Contract
and the other Grants. This refactors the logic to string along a `serviceType`
parameter as the majority of the connection logic, configuration setup, etc.
is the same.

This also tries to abstract the field-copying logic into a `flattenSoapDict`
function, which knows how to deal with "simple" fields (which are just
copied), "address" fields (which must be flattened), and lists which need to
be converted into a comma-separated string. With this, we can just define the
fields in big tuples