Abstract mock file paths to cover more configs
Previously, we had a `mock_file_d_path` function, which modified the
CONFIG_BROKER to use a temporary file path for the `d_file_storage_path`
attribute. The same approach would be helpful for other path configs (starting
with `broker_files`), so this patch abstracts that effort to allow multiple
paths to be replaced with temporary dirs.