Use GlobalDB class for database setup and teardown
At the beginning and end of a request (for both the dataactbroker
and dataactvalidator applications), use the GlobalDB class
to get and tearndown database engines, connections, and sessions.
Note that teardown code is now in teardown_appcontext, which
is run regardless of whether or not there were unexpected errors
in the request (this behavior is not guaranteed when using
after_request).