Add fixes for unicode, non-present attributes
In the process of bulk-downloading the whole corpus of FSRS data, we ran into
some issues that this set of changes should resolve.

First, we were seeing an explosion due to suds/sax viewing certain control
characters as "invalid tokens". To resolve that, this adds a filter to replace
those chars with spaces.

Second, some records didn't have the `subawardees` or `subcontracts` fields.
While these records don't make much sense, they are up to the WSDL spec, so
we'll allow them.