Run validation schema load outside of global database connection.
This is the first step of transition the validator initialization
process away from the global database object. The goal of this
transition is to use the new databaseSession context for all
non-Flask db access. Subsequent steps will transition the
remainder of the validation initialize work: tas load and
domain values load.