Temporarily copy password hashing logic to global db functions.
I'd hoped to set up the global db functions in a way that non-Flask
db logic wouldn't need to use them. However, the password hashing
function isn't something I wanted to repeat throughout the code base.
Ultimately, I think that logic belongs at the model level so that
application code doesn't have to deal with it. For now, copying it
to the beginning of our global db functions list so that the init
scripts and integration tests can access it there. Will refactor
again when updating the Flask db access.