Remove databaseSession context
The team decided that it would be better to have a single,
consistent way to access the database, regardless of whether
or not the code being executed is part of the a flask application.
This changeset replaces the old databaseSession context we were
using for some of the non-Flask db access and replaces it with
code that creates a flask application context for db access.