Change TAS file format
We're getting our TAS files from a different source now, which has a slightly
different set of columns. Modify the import script to accommodate and replace
the CSVs included in the app. The logic is simplified as our data source is
now more accurate (not including duplicate rows, for example).