Create a consolidated broker initialization script
Create a single initialization script for the data act broker.
This is a first pass. The new scripts is more flexible than
previous versions, allowing people to mix and match various
setup tasks. That said, I'm not sure if all of this belongs
in a single script, or if it would be better to have one
script for db setup and one related to data loads.