If SessionTable already exists when creating it, catch exception.
Update the DynmamoDB SessionTable creation function to catch
the exception if SessionTable already exists. Previously, we
hard worked around this by omitting this step from the deploy's
init script, but we decided to make the DynamoDB table setup
consistent with other table setups (i.e., doesn't mind if tables
are already there). This is also how the local Dynamo was working.