Remove getAgencyName function from validationInteface
This is another function that's short and simple enough to
be called inline. Note that at some point in the future, it
might be worth setting up a SQLAlchemy relationship between
user and cgac, so we can pull back a user's agency name without
an additional database call. I left this enhancement out of
scope for the current PR.