Remove the short/long column mapping functions from validationInterface
This commit removes the last two functions form validationInterface.
We now create the short-long column mappings when the ValidationManager
is created. To make this work, I moved the instantiation of ValidationManager
from the validator's createApp method to the beginning of the validate/
route. It seems reasonable to create a new one of these for each
validation job.