Move runChecks from jobTrackerInterface to function_bag
As part of the move, rename the function to make it clearer
that the checks being run pertain to checking a job's eligibility
to run. Although the query logic is simpler now and could, in
theory, be run as single statement as needed, checking a job's
eligibility seemed important enough/subject enough to change
that it seems worth centralizing. Lastly, the function no longer
raises an error so the application code can evaluate the results
and raise errors as necessary.