Fix intermittent Flask AssertionErrors in integration tests
When a Flask app is in debug mode, it's PRESERVE_CONTEXT_ON_EXCEPTION
value is set to True, which ensures that an exception will not
tear down a request immediately (to ensure that the interactive
debugger can still access it). However, having that extra context
around results in an AssertionError: Popped wrong request context
if the teardown method raises an error. Setting PRESERVE_CONTEXT_ON_
EXECPTION to False tells the Flask app to tear down a request on
exception, ensuring that we see the real problem when running
tests, not the issue re: popping the wrong context.