Remove JsonResponse's CloudLogger use
Unfortunately, this requires a lot of refactoring. Given that CloudLogger (and
JsonResponse.error) make heavy use of exceptions, it would seem like we could
assume that their processing was occurring within the exception context (and
hence we could use logging.exception to grab the traceback, etc.
automatically). Instead, we have this odd combination of catching exceptions
in some cases and creating them to pass around in others. To standardize,
this patch switches a bunch of functions to _raising_ exceptions rather than
just creating them. This should be a stop-gap until we have top-level
exception handling (which is outside the scope of this PR).

Along the way, it refactors several functions to use a dictionary to represent
HTTP parameters, switches variables to use underscores, etc. When reviewing,
please double check that I've matched the previous logic.