Add tas_id to staging models for A, B, C
After inserting Appropriations, AwardFinancials, and
ObjectClassProgramActivities, execute some SQL that'll update the entries to
include an FK to a tas_lookup value. This is a step closer to removing the
`tas` concatenated field (as we'll be able to compare `tas_ids`).

I'm not too keen on this raw SQL, but the sqlalchemy equivalent seemed harder
to comprehend and was full of gotchas.