Post JSON when calling the validate route
The code that grabs the job_id at the beginning of the validate
route expects the request data to be formatted as json. This
code passed integration tests, because those tests post json
to the validate route. However, it didn't work in actual use
because the broker code that enqueues the requests for celery
wasn't posting as json. This commit changes that.