Improve validator error handlers
This changeset adds some job cleanup code to the validator's
handler for uncaught exceptions. In addition, it adds a few
checks to the ResponseException handler to ensure that error
handling works correctly in all scenarios (for example, writing
file error information where needed). It's clear that the job
cleanup piece is similar between the two validator error
handlers, so a future commit will refactor that.