Reinstate submission not found error handling
The previous version of check_submission_by_id does some
important error handling when the passed submission doesn't
exist. This commit reinstates that error handling, albeit
without the ability to include a trace. The goal is to strike
a balance between sending the expected response to the front-end,
having readable code on the back end, and expediency. Longer term,
it would ideal to handle some of these not found errors more
globally.