Sort list of files before create jobs and dependencies
This commit makes it more explicit that we need to set up
jobs and dependencies in a particular order (namely, the
jobs for awards files D1 and D2 need to be created prior to
the creation of sub-award and awardee attributes jobs. This
is moot in the long term, since there's a story in the backlog
to eliminate job create for sub-award and awardee jobs, but
while we're still doing it, best to be explicit.