Replace config.py
We originally had one config.py, but we needed to add a log statement to it.
Unfortunately, we were using CloudLogger at the time, which introduced a
cyclic dependency. To break it, we moved config.py to read_config.py and had
both CloudLogger and config.py import read_config.py.

Now that CloudLogger is dead, we don't need two config files.