Remove submission start/end date formatting from jobHandler
Instead of using functions to format a submission's start and
end reporting dates, the check_status route now does that
formatting inline. The logic is simplified by
using the fiscal period/year information that is now included
with each submission. To ensure that the code works on older
submissions, this changeset includes a data migration to fix
up submissions that are missing fiscal period/year information.