Avoid "is not distinct from" in tas_id for SF133s
SQL's "is not distinct from" (which is like equality + treating nulls as a
single value) does not make use of indexes. Therefore, we need to swap out
these lookups with `left == right or (left is null and right is null)`.
Sqlalchemy gives us a way to describe that repetition concisely.