Add website_admin field to User
The website admin permission is different from the other permission types in
that it is not specific to a particular agency. As we'll be allowing a single
user to have different types of permissions for different agencies shortly,
it's a good time to separate out the website_admin permission now.

Migrates existing data as we're removing the lookup