Add level awareness to requires_submission_perms
Replace the requires_submission_perms function (which converted submission_id
to submissions and verified 1) that the submission existing and 2) the user's
agency matched) to a version which is aware of permission levels. This
effectively combines it with `current_user_can_on_submission`.

That changes the signature for the narrative functions, so their tests are
also updated.