Use matching_cars_subquery on staging models
Replaces the hand-crafted SQL with an almost equivalent sqlalchemy call. The
big difference is that the sqlalchemy subquery doesn't use "is not distinct
from", which can't take advantage of indexes.

Also removes the `onupdate` configs from the staging models as we're mass
updating them. We could avoid this before by working around the ORM.